Add a MEE specification
The actual implementation of the spec is odd (this whole
chapter-per-function thing looks bad), but I think it at least gets the
point across for now.  There's very little that's actually usable in the
spec as it currently stands because we don't have a way to detect
hardware devices so users can only call the function that have an
implicit default device.

Signed-off-by: Palmer Dabbelt <palmer@sifive.com>