Stop using watermark. Correct some define directives
` Watermarks do not work because when a FIFO is full, the only control
I have is that the pending bit in IP register will get cleared. I cannot
make use of this information to enqueue severy bytes of data before
sending them all out. The pause and transmit do not seem to be
controlled by the user of the hardware interface.

` Correct the value for txdata watermark mask. Should be 7 instead of 3 because the [2:0] digits represent the watermark,