Ensure TXDATA is written once in each iteration
Previously, I used ^= ~<mask> to mask out the least significant byte of
TXDATA, which is the data that gets enqueued. Then I |= data to actually
write it. In this process, the TXDATA register is written to twice. So
while my tx_buf probably is [0x1, 0x2, 0x3]. The actual data that get
sent out is [0, 0x1, 0, 0x2, 0, 0x3].