Enabled blocking sys_read, added circular buffer
Implemented a circular buffer that is terminated by '\r' or '\n' in
sys_read. Characters will be stored in this buffer before read into
actual storage. This buffer has a hardcoded wrap-around length of 127.
More work needs to be done to deal with possible exceptions because of
this maximum length.

This circular buffer also most closely resembles various I/O functions
when they are executed normally on a PC.

Non-blocking read requires playing with fcntl and the relevant drivers.
Will deal with that in the future.