Prevent the user from typing more than buffer limit
Only allow the user to press enter or backspace or del when the number
of characters in the buffer is TTY_BUFFER_LEN - 1, beucause the last
character will always be '\n'. It is not recommended for the buffer to
approach this limit because the behavior for the last two characters are
usually undefined. Although tty_buffer is declared static global,
metal_tty_putc constantly overrides the memory reagion of tty_buffer
when the length of messages approach TTY_BUFFER_LEN.

start_index and end_index are also affected by calls of metal_tty_putc.
The current workaround is a macro that backups the values of these two
indices on the stack of function present_buffer() before calling
metal_tty_putc(). After metal_tty_putc(), the values of the two indices
are restored.

This tactic is not used to ensure the integrity of tty_buffer[] as
buffer length approaches TTY_BUFFER_LEN, because comparing two arrays
before and after each call of metal_tty_putc() is too costly.