drivers/riscv_plic0: Disable all interrupts on init
The PLIC driver was only disabling interrupts for up to ndev devices,
but there may be more devices on the target because we shrink ndev to
avoid allocating too much memory for interrupt handlers.  This patch
forces the PLIC driver to always disable 1024 interrupts.

We should probably introduce a better scheme for looking up interrupt
handlers, but for now this will at least turn off all the interrupts on
any PLIC that has ever been configured.

I haven't tested this.

Signed-off-by: Palmer Dabbelt <palmer@sifive.com>