Refactor and simplify autotools configuration
Removes all old build system components, namely the parts that allowed
the Freedom Metal build to run dtc and freedom-devicetree-tools on a
devicetree to generate the header files. The result is a build system
that reflects how we actually use the library today, and it's much
easier to see how the build works now.

Signed-off-by: Nathaniel Graff <nathaniel.graff@sifive.com>