Refactor BEU driver for freedom-metal-next
Special-case the sifive,buserror0 local interrupt handler because it has
a hard-coded interrupt ID of 128 which isn't expressed in the
devicetree. For software-vectored interrupts, this means we just don't
include the handler in the local interrupt table. For HLIC
hardware-vectored interrupts, we have no choice but to waste a bunch of
space in the jump table to put the handler at the right offset.

Signed-off-by: Nathaniel Graff <nathaniel.graff@sifive.com>