Generated paths are relative to ROOT_DIR
I would have written a shorter commit, but I didn't
have the time. :)

In order to accomodate paths in freedom-e-sdk, standalone
projects, and the Wake build environment, all paths
generated in metal.mk are relative to the ROOT_DIR make
variable.

We have to use relative paths instead of absolute paths
because the generated paths will be part of the files
included in customer deliverables, and the paths must be
valid both during the Wake build and in the user's
environment.

However, simply setting a relative path won't allow
for standalone project creation, because the program sources
are in a different location relative to Freedom Metal
sources and generated files depending on whether we're
in the SDK or a standalone project.

Therefore, we use a new Make variable ROOT_DIR to tell us
where the root of the SDK or standalone project is and
generate paths relative to it.

The ROOT_DIR variable should be set in freedom-e-sdk
standalone.mk to point at the root directory of the SDK or
standalone project.

Signed-off-by: Nathaniel Graff <nathaniel.graff@sifive.com>